refactor: moved the permissions to the users app
