+++ apps/users/permissions.py
@@ -1,6 +1,6 @@
 from rest_framework.exceptions import PermissionDenied
 from rest_framework.permissions import BasePermission, SAFE_METHODS
-
+from apps.users.models import Roles
 
 class ReadOnly(BasePermission):
     def has_permission(self, request, view):
@@ -51,3 +51,24 @@ class IsSuperAdmin(BasePermission):
 class NotAllowed(BasePermission):
     def has_permission(self, request, view):
         raise PermissionDenied()
+
+class DoctorOrReadOnly(BasePermission):
+    """
+    Custom permission to only allow doctors to edit their availability.
+    """
+
+    def has_permission(self, request, view):
+        # Allow all authenticated users to view doctor availabilities
+        if request.method in SAFE_METHODS:
+            return True
+
+        # Only allow doctors to create availabilities
+        return request.user.role == Roles.DOCTOR
+
+    def has_object_permission(self, request, view, obj):
+        # Allow all authenticated users to view doctor availabilities
+        if request.method in SAFE_METHODS:
+            return True
+
+        # Only allow the specific doctor to update or delete their own availability
+        return obj.doctor == request.user
